Only delete remote files when necessary with theme watch command
Because files that are present locally overwrite remote files anyway,
it is not necessary to delete all files on every execution of `theme
watch`.

Only files that are not present locally are deleted on the remote side.
Speeds up the command extraordinarily in projects with tons of assets
(-> nearly all projects).