Moves over to filewatcher for triggering actions
The listen gem has had this issue for a while and
looking into it requires possibly making patches to
the listen library. Switching to filewatcher requires
no changes except for changing to another library.

Looking through the filewatcher code it doesn't appear
to have any OS specific dependencies, so this may even
fix the issue with users needing to install WDM on top
of this gem when using Windows.