Use a better way of detecting file types
Before we'd have a white-list of known files, but the number of files
that are binary is way bigger than this. Instead of trying to be
exhuastive, let's use a tool that is up to date with all the latest
mime-types and based 'textiness' on that. This requires adding a
special entry for liquid as text files.