Adds configuration checking to commands
In order to provide more details to users and give better
failure modes, we verify the validity of a configuration.

If the configuration is not valid we let the user know and
reference them to the shopify_theme configuration page.

We also check to see if the API is down, and if it is, let the
user know and reference them to the shopify status page.