Replaces file filtering to be more testable
local_assets_list was not the most easy to understand piece of
code, (even though it was pretty small). Instead move over to using
a series of filters and feeding the results into the next filter.

This should making adding more filters easier (if we need those) and
also clarify what our expectations are for this feature.

Fixes Issue #156